novas classes

+++ DashboardParceiro/Models/Cadastros/Categoria.cs
@@ -10,8 +10,7 @@ namespace DashboardParceiro.Models.Cadastros
         [Key]
         [Column("Id")]
         [DisplayName("Codigo")]
-
-        public long Codigo { get; set; }
+        public int Codigo { get; set; }
         [Column("Descricao")]
         [DisplayName("Descrição")]
         public string Descricao { get; set; }
